Consistenly use uint16 for ports in diego
[#86337946]

+++ apiresources.go
@@ -158,8 +158,8 @@ type LogConfig struct {
 }
 
 type PortMapping struct {
-	ContainerPort uint32 `json:"container_port"`
-	HostPort      uint32 `json:"host_port,omitempty"`
+	ContainerPort uint16 `json:"container_port"`
+	HostPort      uint16 `json:"host_port,omitempty"`
 }
 
 type ContainerRunResult struct {
